+++ CHANGELOG.md
@@ -6,9 +6,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [0.1.18] - 2020-12-24
+## [0.1.20] - 2020-12-24
 ### Added
 - C#\DateTime.cs
+- C#\Select. Example 2.cs
+- Webinars\REST tools.txt
 
 ## [0.1.17] - 2020-12-23
 ### Changed